refine interval timer for scalar
